refactor(core): split install module responsibilities
