--- crates/surge-core/src/install.rs
@@ -1,942 +1,0 @@
-use std::collections::BTreeMap;
-use std::path::{Component, Path, PathBuf};
-use std::time::Duration;
-
-use serde::Serialize;
-use tracing::warn;
-
-use crate::archive::extractor::extract_file_to_with_progress;
-use crate::config::installer::InstallerManifest;
-use crate::config::manifest::ShortcutLocation;
-use crate::context::StorageProvider;
-use crate::error::{Result, SurgeError};
-use crate::platform::fs::{copy_directory, list_directories};
-use crate::platform::paths::default_install_root;
-use crate::platform::process::{ProcessHandle, spawn_detached};
-use crate::platform::shortcuts::install_shortcuts;
-use crate::releases::version::compare_versions;
-use crate::supervisor::stub::find_latest_app_dir;
-
-pub const RUNTIME_MANIFEST_RELATIVE_PATH: &str = ".surge/runtime.yml";
-pub const LEGACY_RUNTIME_MANIFEST_RELATIVE_PATH: &str = ".surge/surge.yml";
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum InstallProgressStage {
-    Extract,
-    Activate,
-    Shortcuts,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub struct InstallProgress {
-    pub stage: InstallProgressStage,
-    pub phase_percent: i32,
-    pub bytes_done: i64,
-    pub bytes_total: i64,
-    pub items_done: i64,
-    pub items_total: i64,
-}
-
-pub type InstallProgressCallback<'a> = dyn Fn(InstallProgress) + Send + Sync + 'a;
-
-fn emit_install_progress(progress: Option<&InstallProgressCallback<'_>>, snapshot: InstallProgress) {
-    if let Some(cb) = progress {
-        cb(snapshot);
-    }
-}
-
-#[derive(Debug, Clone, Default)]
-struct RuntimeManifestSnapshot {
-    runtime_manifest: Option<Vec<u8>>,
-    legacy_runtime_manifest: Option<Vec<u8>>,
-}
-
-impl RuntimeManifestSnapshot {
-    fn capture(active_app_dir: &Path) -> Result<Self> {
-        Ok(Self {
-            runtime_manifest: read_optional_file(&active_app_dir.join(RUNTIME_MANIFEST_RELATIVE_PATH))?,
-            legacy_runtime_manifest: read_optional_file(&active_app_dir.join(LEGACY_RUNTIME_MANIFEST_RELATIVE_PATH))?,
-        })
-    }
-
-    fn restore(&self, active_app_dir: &Path) -> Result<()> {
-        restore_optional_file(
-            &active_app_dir.join(RUNTIME_MANIFEST_RELATIVE_PATH),
-            self.runtime_manifest.as_deref(),
-        )?;
-        restore_optional_file(
-            &active_app_dir.join(LEGACY_RUNTIME_MANIFEST_RELATIVE_PATH),
-            self.legacy_runtime_manifest.as_deref(),
-        )
-    }
-}
-
-fn read_optional_file(path: &Path) -> Result<Option<Vec<u8>>> {
-    if path.is_file() {
-        Ok(Some(std::fs::read(path)?))
-    } else {
-        Ok(None)
-    }
-}
-
-fn restore_optional_file(path: &Path, contents: Option<&[u8]>) -> Result<()> {
-    if let Some(contents) = contents {
-        if let Some(parent) = path.parent() {
-            std::fs::create_dir_all(parent)?;
-        }
-        std::fs::write(path, contents)?;
-    } else if path.exists() {
-        if path.is_dir() {
-            std::fs::remove_dir_all(path)?;
-        } else {
-            std::fs::remove_file(path)?;
-        }
-    }
-
-    Ok(())
-}
-
-/// Shared profile for installing a package locally, usable from both
-/// `surge install` (via `ReleaseEntry`) and `surge setup` (via `InstallerRuntime`).
-pub struct InstallProfile<'a> {
-    pub app_id: &'a str,
-    pub display_name: &'a str,
-    pub main_exe: &'a str,
-    pub install_directory: &'a str,
-    pub supervisor_id: &'a str,
-    pub icon: &'a str,
-    pub shortcuts: &'a [ShortcutLocation],
-    pub persistent_assets: &'a [String],
-    pub environment: &'a BTreeMap<String, String>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct RuntimeManifestMetadata<'a> {
-    pub version: &'a str,
-    pub channel: &'a str,
-    pub storage_provider: &'a str,
-    pub storage_bucket: &'a str,
-    pub storage_region: &'a str,
-    pub storage_endpoint: &'a str,
-}
-
-impl<'a> RuntimeManifestMetadata<'a> {
-    #[must_use]
-    pub fn new(
-        version: &'a str,
-        channel: &'a str,
-        storage_provider: &'a str,
-        storage_bucket: &'a str,
-        storage_region: &'a str,
-        storage_endpoint: &'a str,
-    ) -> Self {
-        Self {
-            version,
-            channel,
-            storage_provider,
-            storage_bucket,
-            storage_region,
-            storage_endpoint,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize)]
-struct RuntimeManifestFile<'a> {
-    id: &'a str,
-    version: &'a str,
-    channel: &'a str,
-    #[serde(rename = "installDirectory")]
-    install_directory: &'a str,
-    #[serde(rename = "supervisorId", skip_serializing_if = "str::is_empty")]
-    supervisor_id: &'a str,
-    provider: &'a str,
-    bucket: &'a str,
-    #[serde(skip_serializing_if = "str::is_empty")]
-    region: &'a str,
-    #[serde(skip_serializing_if = "str::is_empty")]
-    endpoint: &'a str,
-}
-
-impl<'a> InstallProfile<'a> {
-    #[must_use]
-    pub fn new(
-        app_id: &'a str,
-        display_name: &'a str,
-        main_exe: &'a str,
-        install_directory: &'a str,
-        supervisor_id: &'a str,
-        icon: &'a str,
-        shortcuts: &'a [ShortcutLocation],
-        persistent_assets: &'a [String],
-        environment: &'a BTreeMap<String, String>,
-    ) -> Self {
-        Self {
-            app_id,
-            display_name,
-            main_exe,
-            install_directory,
-            supervisor_id,
-            icon,
-            shortcuts,
-            persistent_assets,
-            environment,
-        }
-    }
-
-    #[must_use]
-    pub fn from_installer_manifest(manifest: &'a InstallerManifest, shortcuts: &'a [ShortcutLocation]) -> Self {
-        Self::new(
-            &manifest.app_id,
-            &manifest.runtime.name,
-            &manifest.runtime.main_exe,
-            &manifest.runtime.install_directory,
-            &manifest.runtime.supervisor_id,
-            &manifest.runtime.icon,
-            shortcuts,
-            &manifest.runtime.persistent_assets,
-            &manifest.runtime.environment,
-        )
-    }
-}
-
-#[must_use]
-pub fn storage_provider_manifest_name(provider: Option<StorageProvider>) -> &'static str {
-    match provider.unwrap_or(StorageProvider::Filesystem) {
-        StorageProvider::S3 => "s3",
-        StorageProvider::AzureBlob => "azure",
-        StorageProvider::Gcs => "gcs",
-        StorageProvider::Filesystem => "filesystem",
-        StorageProvider::GitHubReleases => "github_releases",
-    }
-}
-
-pub fn write_runtime_manifest(
-    active_app_dir: &Path,
-    profile: &InstallProfile<'_>,
-    metadata: &RuntimeManifestMetadata<'_>,
-) -> Result<std::path::PathBuf> {
-    let manifest = RuntimeManifestFile {
-        id: profile.app_id.trim(),
-        version: metadata.version.trim(),
-        channel: metadata.channel.trim(),
-        install_directory: profile.install_directory.trim(),
-        supervisor_id: profile.supervisor_id.trim(),
-        provider: metadata.storage_provider.trim(),
-        bucket: metadata.storage_bucket.trim(),
-        region: metadata.storage_region.trim(),
-        endpoint: metadata.storage_endpoint.trim(),
-    };
-
-    if manifest.id.is_empty() {
-        return Err(SurgeError::Config(
-            "Cannot write runtime manifest: app id is empty".to_string(),
-        ));
-    }
-    if manifest.version.is_empty() {
-        return Err(SurgeError::Config(
-            "Cannot write runtime manifest: version is empty".to_string(),
-        ));
-    }
-    if manifest.channel.is_empty() {
-        return Err(SurgeError::Config(
-            "Cannot write runtime manifest: channel is empty".to_string(),
-        ));
-    }
-    if manifest.provider.is_empty() {
-        return Err(SurgeError::Config(
-            "Cannot write runtime manifest: storage provider is empty".to_string(),
-        ));
-    }
-
-    let runtime_manifest_path = active_app_dir.join(RUNTIME_MANIFEST_RELATIVE_PATH);
-    if let Some(parent) = runtime_manifest_path.parent() {
-        std::fs::create_dir_all(parent)?;
-    }
-    let legacy_manifest_path = active_app_dir.join(LEGACY_RUNTIME_MANIFEST_RELATIVE_PATH);
-    if let Some(parent) = legacy_manifest_path.parent() {
-        std::fs::create_dir_all(parent)?;
-    }
-
-    let mut yaml = serde_yaml::to_string(&manifest)
-        .map_err(|e| SurgeError::Config(format!("Failed to serialize runtime manifest: {e}")))?;
-    if !yaml.ends_with('\n') {
-        yaml.push('\n');
-    }
-    std::fs::write(&runtime_manifest_path, &yaml)?;
-    std::fs::write(&legacy_manifest_path, yaml)?;
-    Ok(runtime_manifest_path)
-}
-
-/// Resolve the install root and install the package there.
-pub fn install_package_locally(profile: &InstallProfile<'_>, package_path: &Path) -> Result<std::path::PathBuf> {
-    let install_root = default_install_root(profile.app_id, profile.install_directory)?;
-    install_package_locally_at_root(profile, package_path, &install_root)?;
-    Ok(install_root)
-}
-
-/// Extract a package into `install_root/app` with atomic swap, then create shortcuts.
-pub fn install_package_locally_at_root(
-    profile: &InstallProfile<'_>,
-    package_path: &Path,
-    install_root: &Path,
-) -> Result<()> {
-    install_package_locally_at_root_with_progress(profile, package_path, install_root, None)
-}
-
-/// Extract a package into `install_root/app` with atomic swap, then create shortcuts.
-pub fn install_package_locally_at_root_with_progress(
-    profile: &InstallProfile<'_>,
-    package_path: &Path,
-    install_root: &Path,
-    progress: Option<&InstallProgressCallback<'_>>,
-) -> Result<()> {
-    std::fs::create_dir_all(install_root)?;
-
-    let active_app_dir = install_root.join("app");
-    let next_app_dir = install_root.join(".surge-app-next");
-    let previous_app_dir = install_root.join(".surge-app-prev");
-    let fallback_previous_app_dir = if active_app_dir.is_dir() {
-        None
-    } else {
-        find_latest_app_dir(install_root).ok()
-    };
-
-    if next_app_dir.is_dir() {
-        std::fs::remove_dir_all(&next_app_dir)?;
-    }
-    if previous_app_dir.is_dir() {
-        std::fs::remove_dir_all(&previous_app_dir)?;
-    }
-
-    emit_install_progress(
-        progress,
-        InstallProgress {
-            stage: InstallProgressStage::Extract,
-            phase_percent: 0,
-            bytes_done: 0,
-            bytes_total: 0,
-            items_done: 0,
-            items_total: 0,
-        },
-    );
-    let extract_progress = |items_done: u64, items_total: u64, bytes_done: u64, bytes_total: u64| {
-        let phase_percent = if bytes_total > 0 {
-            ((bytes_done.saturating_mul(100)) / bytes_total).clamp(0, 100) as i32
-        } else if items_total > 0 {
-            ((items_done.saturating_mul(100)) / items_total).clamp(0, 100) as i32
-        } else {
-            0
-        };
-        emit_install_progress(
-            progress,
-            InstallProgress {
-                stage: InstallProgressStage::Extract,
-                phase_percent,
-                bytes_done: i64::try_from(bytes_done).unwrap_or(i64::MAX),
-                bytes_total: i64::try_from(bytes_total).unwrap_or(i64::MAX),
-                items_done: i64::try_from(items_done).unwrap_or(i64::MAX),
-                items_total: i64::try_from(items_total).unwrap_or(i64::MAX),
-            },
-        );
-    };
-    extract_file_to_with_progress(package_path, &next_app_dir, Some(&extract_progress))?;
-    emit_install_progress(
-        progress,
-        InstallProgress {
-            stage: InstallProgressStage::Extract,
-            phase_percent: 100,
-            bytes_done: 0,
-            bytes_total: 0,
-            items_done: 0,
-            items_total: 0,
-        },
-    );
-
-    emit_install_progress(
-        progress,
-        InstallProgress {
-            stage: InstallProgressStage::Activate,
-            phase_percent: 0,
-            bytes_done: 0,
-            bytes_total: 0,
-            items_done: 0,
-            items_total: 0,
-        },
-    );
-    if active_app_dir.is_dir() {
-        std::fs::rename(&active_app_dir, &previous_app_dir)?;
-    }
-
-    if let Err(rename_err) = std::fs::rename(&next_app_dir, &active_app_dir) {
-        if previous_app_dir.is_dir() && !active_app_dir.exists() {
-            let _ = std::fs::rename(&previous_app_dir, &active_app_dir);
-        }
-        return Err(SurgeError::Io(rename_err));
-    }
-    emit_install_progress(
-        progress,
-        InstallProgress {
-            stage: InstallProgressStage::Activate,
-            phase_percent: 100,
-            bytes_done: 0,
-            bytes_total: 0,
-            items_done: 0,
-            items_total: 0,
-        },
-    );
-
-    let runtime_manifest_snapshot = RuntimeManifestSnapshot::capture(&active_app_dir)?;
-    let install_result = (|| -> Result<()> {
-        let previous_app_dir_for_assets = if previous_app_dir.is_dir() {
-            Some(previous_app_dir.as_path())
-        } else {
-            fallback_previous_app_dir.as_deref()
-        };
-        if !profile.persistent_assets.is_empty()
-            && let Some(previous) = previous_app_dir_for_assets
-        {
-            copy_persistent_assets(previous, &active_app_dir, profile.persistent_assets)?;
-        }
-
-        if !profile.shortcuts.is_empty() {
-            emit_install_progress(
-                progress,
-                InstallProgress {
-                    stage: InstallProgressStage::Shortcuts,
-                    phase_percent: 0,
-                    bytes_done: 0,
-                    bytes_total: 0,
-                    items_done: 0,
-                    items_total: 0,
-                },
-            );
-            let main_exe = profile.main_exe.trim();
-            if main_exe.is_empty() {
-                return Err(SurgeError::Config(format!(
-                    "App '{}' has shortcuts configured but no main executable metadata",
-                    profile.app_id
-                )));
-            }
-            install_shortcuts(
-                profile.app_id,
-                profile.display_name,
-                &active_app_dir,
-                main_exe,
-                profile.supervisor_id,
-                profile.icon,
-                profile.shortcuts,
-                profile.environment,
-            )?;
-            emit_install_progress(
-                progress,
-                InstallProgress {
-                    stage: InstallProgressStage::Shortcuts,
-                    phase_percent: 100,
-                    bytes_done: 0,
-                    bytes_total: 0,
-                    items_done: i64::try_from(profile.shortcuts.len()).unwrap_or(i64::MAX),
-                    items_total: i64::try_from(profile.shortcuts.len()).unwrap_or(i64::MAX),
-                },
-            );
-        }
-
-        if previous_app_dir.is_dir() {
-            std::fs::remove_dir_all(&previous_app_dir)?;
-        }
-
-        Ok(())
-    })();
-    if let Err(err) = install_result {
-        if let Err(restore_err) = runtime_manifest_snapshot.restore(&active_app_dir) {
-            return Err(SurgeError::Platform(format!(
-                "Failed to restore runtime manifests after install error '{err}': {restore_err}"
-            )));
-        }
-        return Err(err);
-    }
-
-    if let Err(err) = prune_version_snapshots(install_root, 0) {
-        warn!(error = %err, "Failed to prune installed app version snapshots after install");
-    }
-
-    Ok(())
-}
-
-fn app_snapshot_version(dir_name: &str) -> Option<&str> {
-    let version = dir_name.strip_prefix("app-")?;
-    if version.is_empty() || !version.chars().next().is_some_and(|c| c.is_ascii_digit()) {
-        return None;
-    }
-    Some(version)
-}
-
-pub fn prune_version_snapshots(install_dir: &Path, keep_latest: usize) -> Result<usize> {
-    let mut snapshots: Vec<(String, PathBuf)> = list_directories(install_dir)?
-        .into_iter()
-        .filter_map(|dir_name| {
-            let version = app_snapshot_version(&dir_name)?.to_string();
-            let path = install_dir.join(&dir_name);
-            Some((version, path))
-        })
-        .collect();
-
-    snapshots.sort_by(|(left, _), (right, _)| compare_versions(right, left));
-
-    let mut pruned = 0usize;
-    for (_, path) in snapshots.into_iter().skip(keep_latest) {
-        std::fs::remove_dir_all(path)?;
-        pruned = pruned.saturating_add(1);
-    }
-
-    Ok(pruned)
-}
-
-pub fn copy_persistent_assets(previous_app_dir: &Path, new_app_dir: &Path, assets: &[String]) -> Result<()> {
-    for asset in assets {
-        let relative = validate_relative_persistent_asset_path(asset)?;
-        let source = previous_app_dir.join(&relative);
-        if !source.exists() {
-            continue;
-        }
-
-        let destination = new_app_dir.join(&relative);
-        if source.is_dir() {
-            if destination.exists() {
-                if destination.is_dir() {
-                    std::fs::remove_dir_all(&destination)?;
-                } else {
-                    std::fs::remove_file(&destination)?;
-                }
-            }
-            copy_directory(&source, &destination)?;
-        } else {
-            if let Some(parent) = destination.parent() {
-                std::fs::create_dir_all(parent)?;
-            }
-            if destination.exists() && destination.is_dir() {
-                std::fs::remove_dir_all(&destination)?;
-            }
-            std::fs::copy(&source, &destination)?;
-        }
-    }
-
-    Ok(())
-}
-
-pub fn validate_relative_persistent_asset_path(raw: &str) -> Result<PathBuf> {
-    if raw.trim().is_empty() {
-        return Err(SurgeError::Update("Persistent asset path cannot be empty".to_string()));
-    }
-
-    let candidate = Path::new(raw);
-    if candidate.is_absolute() {
-        return Err(SurgeError::Update(format!(
-            "Persistent asset path must be relative: {raw}"
-        )));
-    }
-
-    let first_component = candidate
-        .components()
-        .next()
-        .and_then(|component| match component {
-            Component::Normal(value) => value.to_str(),
-            _ => None,
-        })
-        .unwrap_or_default();
-    if first_component.to_ascii_lowercase().starts_with("app-") {
-        return Err(SurgeError::Update(format!(
-            "Persistent asset path cannot start with 'app-': {raw}"
-        )));
-    }
-
-    for component in candidate.components() {
-        if matches!(
-            component,
-            Component::ParentDir | Component::RootDir | Component::Prefix(_)
-        ) {
-            return Err(SurgeError::Update(format!(
-                "Persistent asset path cannot contain parent/root traversal: {raw}"
-            )));
-        }
-    }
-
-    Ok(candidate.to_path_buf())
-}
-
-/// Start the installed application, using the supervisor if configured.
-pub fn auto_start_after_install(
-    profile: &InstallProfile<'_>,
-    install_root: &Path,
-    active_app_dir: &Path,
-) -> Result<u32> {
-    start_installed_application(
-        profile,
-        install_root,
-        active_app_dir,
-        &["--surge-installed"],
-        true,
-        false,
-    )
-}
-
-/// Run the post-install lifecycle sequence:
-/// 1) invoke `--surge-installed <version>` and wait for it to finish (up to 15s),
-/// 2) invoke `--surge-first-run <version>` and return that process id.
-///
-/// This mirrors Snapx's install behavior where install hooks run in a short-lived
-/// process and first-run continues in a second process.
-pub fn auto_start_after_install_sequence(
-    profile: &InstallProfile<'_>,
-    install_root: &Path,
-    active_app_dir: &Path,
-    version: &str,
-) -> Result<u32> {
-    let installed_args = lifecycle_args("--surge-installed", version);
-    let mut installed_handle =
-        spawn_installed_application(profile, install_root, active_app_dir, &installed_args, false)?;
-    wait_for_process_exit_or_timeout(&mut installed_handle, Duration::from_secs(15));
-
-    let first_run_args = lifecycle_args("--surge-first-run", version);
-    let first_run_handle = spawn_installed_application(profile, install_root, active_app_dir, &first_run_args, false)?;
-    Ok(first_run_handle.pid())
-}
-
-/// Launch the installed application for user-facing "Launch" actions.
-///
-/// Unlike `auto_start_after_install`, this does not pass the `--surge-installed`
-/// lifecycle argument and should keep GUI apps open for immediate use.
-pub fn launch_installed_application(
-    profile: &InstallProfile<'_>,
-    install_root: &Path,
-    active_app_dir: &Path,
-) -> Result<u32> {
-    start_installed_application(profile, install_root, active_app_dir, &[], false, true)
-}
-
-fn start_installed_application(
-    profile: &InstallProfile<'_>,
-    install_root: &Path,
-    active_app_dir: &Path,
-    app_args: &[&str],
-    prefer_supervisor: bool,
-    verify_running: bool,
-) -> Result<u32> {
-    let mut handle = spawn_installed_application(profile, install_root, active_app_dir, app_args, prefer_supervisor)?;
-    let pid = handle.pid();
-    if verify_running {
-        let process_label = if prefer_supervisor && !profile.supervisor_id.trim().is_empty() {
-            "supervisor"
-        } else {
-            "application"
-        };
-        verify_process_stays_running(&mut handle, process_label)?;
-    }
-    Ok(pid)
-}
-
-fn spawn_installed_application(
-    profile: &InstallProfile<'_>,
-    install_root: &Path,
-    active_app_dir: &Path,
-    app_args: &[&str],
-    prefer_supervisor: bool,
-) -> Result<ProcessHandle> {
-    let main_exe = profile.main_exe.trim();
-    if main_exe.is_empty() {
-        return Err(SurgeError::Config(
-            "Cannot auto-start: no main executable in release metadata".to_string(),
-        ));
-    }
-
-    let exe_path = active_app_dir.join(main_exe);
-
-    let supervisor_id = profile.supervisor_id.trim();
-    if prefer_supervisor && !supervisor_id.is_empty() {
-        let supervisor_path = active_app_dir.join(crate::platform::process::supervisor_binary_name());
-
-        let install_root_str = install_root.to_string_lossy();
-        let exe_path_str = exe_path.to_string_lossy();
-        let mut args: Vec<&str> = vec![
-            "run",
-            "--id",
-            supervisor_id,
-            "--dir",
-            &install_root_str,
-            "--exe",
-            &exe_path_str,
-        ];
-        if !app_args.is_empty() {
-            args.push("--");
-            args.extend_from_slice(app_args);
-        }
-        return spawn_detached(&supervisor_path, &args, Some(install_root), profile.environment);
-    }
-
-    spawn_detached(&exe_path, app_args, Some(install_root), profile.environment)
-}
-
-fn lifecycle_args<'a>(flag: &'a str, version: &'a str) -> Vec<&'a str> {
-    let mut args = vec![flag];
-    let version = version.trim();
-    if !version.is_empty() {
-        args.push(version);
-    }
-    args
-}
-
-fn wait_for_process_exit_or_timeout(handle: &mut ProcessHandle, timeout: Duration) {
-    let check_interval = Duration::from_millis(100);
-    let checks = (timeout.as_millis() / check_interval.as_millis()) as usize;
-
-    for _ in 0..checks {
-        if !handle.poll_running() {
-            let _ = handle.wait();
-            return;
-        }
-        std::thread::sleep(check_interval);
-    }
-
-    if handle.poll_running() {
-        let _ = handle.kill();
-        let _ = handle.wait();
-    }
-}
-
-fn verify_process_stays_running(
-    handle: &mut crate::platform::process::ProcessHandle,
-    process_label: &str,
-) -> Result<()> {
-    let check_interval = Duration::from_millis(200);
-    let total_wait = Duration::from_secs(4);
-    let checks = (total_wait.as_millis() / check_interval.as_millis()) as usize;
-
-    for _ in 0..checks {
-        std::thread::sleep(check_interval);
-        if !handle.poll_running() {
-            let result = handle.wait()?;
-            return Err(SurgeError::Platform(format!(
-                "Failed to launch {process_label}: process exited shortly after start with code {}",
-                result.exit_code
-            )));
-        }
-    }
-    Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::archive::packer::ArchivePacker;
-
-    #[test]
-    fn storage_provider_manifest_name_maps_expected_values() {
-        assert_eq!(storage_provider_manifest_name(None), "filesystem");
-        assert_eq!(storage_provider_manifest_name(Some(StorageProvider::S3)), "s3");
-        assert_eq!(
-            storage_provider_manifest_name(Some(StorageProvider::AzureBlob)),
-            "azure"
-        );
-        assert_eq!(storage_provider_manifest_name(Some(StorageProvider::Gcs)), "gcs");
-        assert_eq!(
-            storage_provider_manifest_name(Some(StorageProvider::Filesystem)),
-            "filesystem"
-        );
-        assert_eq!(
-            storage_provider_manifest_name(Some(StorageProvider::GitHubReleases)),
-            "github_releases"
-        );
-    }
-
-    #[test]
-    fn write_runtime_manifest_creates_expected_file() {
-        let tmp = tempfile::tempdir().expect("temp dir should exist");
-        let environment = BTreeMap::new();
-        let shortcuts: [ShortcutLocation; 0] = [];
-        let persistent_assets: [String; 0] = [];
-        let profile = InstallProfile::new(
-            "demo-app",
-            "Demo App",
-            "demo",
-            "demo-install",
-            "demo-supervisor",
-            "",
-            &shortcuts,
-            &persistent_assets,
-            &environment,
-        );
-        let metadata =
-            RuntimeManifestMetadata::new("1.2.3", "test", "azure", "demo-bucket", "", "https://example.invalid");
-
-        let path = write_runtime_manifest(tmp.path(), &profile, &metadata).expect("runtime manifest should be written");
-        let raw = std::fs::read_to_string(&path).expect("runtime manifest should be readable");
-        let legacy_raw = std::fs::read_to_string(tmp.path().join(LEGACY_RUNTIME_MANIFEST_RELATIVE_PATH))
-            .expect("legacy runtime manifest should be readable");
-        assert!(raw.contains("id: demo-app"));
-        assert!(raw.contains("version: 1.2.3"));
-        assert!(raw.contains("channel: test"));
-        assert!(raw.contains("installDirectory: demo-install"));
-        assert!(raw.contains("supervisorId: demo-supervisor"));
-        assert!(raw.contains("provider: azure"));
-        assert!(raw.contains("bucket: demo-bucket"));
-        assert!(raw.contains("endpoint: https://example.invalid"));
-        assert!(!raw.contains("region:"));
-        assert_eq!(raw, legacy_raw);
-    }
-
-    #[test]
-    fn install_package_locally_preserves_declared_persistent_assets_and_prunes_snapshots() {
-        let tmp = tempfile::tempdir().expect("temp dir should exist");
-        let install_root = tmp.path().join("install-root");
-        let active_app_dir = install_root.join("app");
-        let package_path = tmp.path().join("package.tar.zst");
-        let environment = BTreeMap::new();
-        let shortcuts: [ShortcutLocation; 0] = [];
-        let persistent_assets = vec!["settings.json".to_string(), "state".to_string()];
-
-        std::fs::create_dir_all(active_app_dir.join("state")).expect("state dir should exist");
-        std::fs::create_dir_all(install_root.join("app-1.0.0")).expect("snapshot should exist");
-        std::fs::create_dir_all(install_root.join("app-0.9.0")).expect("snapshot should exist");
-        std::fs::write(active_app_dir.join("settings.json"), "persisted settings").expect("settings should exist");
-        std::fs::write(active_app_dir.join("state").join("cache.bin"), "persisted cache").expect("state should exist");
-        std::fs::write(active_app_dir.join("old.txt"), "remove me").expect("old file should exist");
-
-        let mut packer = ArchivePacker::new(3).expect("archive packer should be created");
-        packer
-            .add_buffer("demo", b"#!/bin/sh\necho demo\n", 0o755)
-            .expect("main executable should be added");
-        packer
-            .add_buffer("settings.json", b"packaged settings", 0o644)
-            .expect("settings should be added");
-        packer
-            .add_buffer("state/cache.bin", b"packaged cache", 0o644)
-            .expect("state should be added");
-        packer
-            .add_buffer("payload.txt", b"new payload", 0o644)
-            .expect("payload should be added");
-        packer
-            .finalize_to_file(&package_path)
-            .expect("archive should be written");
-
-        let profile = InstallProfile::new(
-            "demo-app",
-            "Demo App",
-            "demo",
-            "demo-install",
-            "",
-            "",
-            &shortcuts,
-            &persistent_assets,
-            &environment,
-        );
-
-        install_package_locally_at_root(&profile, &package_path, &install_root).expect("install should succeed");
-
-        let active_app_dir = install_root.join("app");
-        assert_eq!(
-            std::fs::read_to_string(active_app_dir.join("settings.json")).expect("settings should exist"),
-            "persisted settings"
-        );
-        assert_eq!(
-            std::fs::read_to_string(active_app_dir.join("state").join("cache.bin")).expect("state should exist"),
-            "persisted cache"
-        );
-        assert_eq!(
-            std::fs::read_to_string(active_app_dir.join("payload.txt")).expect("payload should exist"),
-            "new payload"
-        );
-        assert!(
-            !active_app_dir.join("old.txt").exists(),
-            "undeclared assets should be removed"
-        );
-        assert!(
-            !install_root.join("app-1.0.0").exists(),
-            "stale snapshot should be pruned"
-        );
-        assert!(
-            !install_root.join("app-0.9.0").exists(),
-            "stale snapshot should be pruned"
-        );
-        assert!(
-            !install_root.join(".surge-app-prev").exists(),
-            "temporary previous dir should be removed"
-        );
-    }
-
-    #[test]
-    fn install_package_locally_restores_runtime_manifests_when_post_copy_work_fails() {
-        let tmp = tempfile::tempdir().expect("temp dir should exist");
-        let install_root = tmp.path().join("install-root");
-        let active_app_dir = install_root.join("app");
-        let package_path = tmp.path().join("package.tar.zst");
-        let environment = BTreeMap::new();
-        let shortcuts = [ShortcutLocation::Desktop];
-        let persistent_assets = vec![".surge".to_string()];
-        let old_runtime_manifest = "id: demo-app\nversion: 1.0.0\nchannel: stable\n";
-        let new_runtime_manifest = "id: demo-app\nversion: 1.1.0\nchannel: beta\n";
-
-        std::fs::create_dir_all(active_app_dir.join(".surge")).expect(".surge dir should exist");
-        std::fs::write(
-            active_app_dir.join(RUNTIME_MANIFEST_RELATIVE_PATH),
-            old_runtime_manifest,
-        )
-        .expect("old runtime manifest should exist");
-        std::fs::write(
-            active_app_dir.join(LEGACY_RUNTIME_MANIFEST_RELATIVE_PATH),
-            old_runtime_manifest,
-        )
-        .expect("old legacy runtime manifest should exist");
-        std::fs::write(active_app_dir.join("payload.txt"), "old payload").expect("old payload should exist");
-
-        let mut packer = ArchivePacker::new(3).expect("archive packer should be created");
-        packer
-            .add_buffer(RUNTIME_MANIFEST_RELATIVE_PATH, new_runtime_manifest.as_bytes(), 0o644)
-            .expect("runtime manifest should be added");
-        packer
-            .add_buffer(
-                LEGACY_RUNTIME_MANIFEST_RELATIVE_PATH,
-                new_runtime_manifest.as_bytes(),
-                0o644,
-            )
-            .expect("legacy runtime manifest should be added");
-        packer
-            .add_buffer("payload.txt", b"new payload", 0o644)
-            .expect("payload should be added");
-        packer
-            .finalize_to_file(&package_path)
-            .expect("archive should be written");
-
-        let profile = InstallProfile::new(
-            "demo-app",
-            "Demo App",
-            "",
-            "demo-install",
-            "",
-            "",
-            &shortcuts,
-            &persistent_assets,
-            &environment,
-        );
-
-        let err = install_package_locally_at_root(&profile, &package_path, &install_root)
-            .expect_err("install should fail before shortcut creation");
-        assert!(
-            matches!(err, SurgeError::Config(ref message) if message.contains("shortcuts configured")),
-            "unexpected error: {err}"
-        );
-        assert_eq!(
-            std::fs::read_to_string(active_app_dir.join(RUNTIME_MANIFEST_RELATIVE_PATH))
-                .expect("runtime manifest should exist"),
-            new_runtime_manifest
-        );
-        assert_eq!(
-            std::fs::read_to_string(active_app_dir.join(LEGACY_RUNTIME_MANIFEST_RELATIVE_PATH))
-                .expect("legacy runtime manifest should exist"),
-            new_runtime_manifest
-        );
-        assert_eq!(
-            std::fs::read_to_string(active_app_dir.join("payload.txt")).expect("payload should exist"),
-            "new payload"
-        );
-    }
-}
