+++ docs/architecture/maintainability-baseline.txt
@@ -5,7 +5,6 @@
 813 crates/surge-cli/src/commands/install/mod.rs
 1860 crates/surge-cli/src/commands/install/remote.rs
 835 crates/surge-cli/src/main.rs
-727 crates/surge-core/src/install.rs
 782 crates/surge-core/src/releases/delta.rs
 604 crates/surge-core/src/storage/azure.rs
 840 crates/surge-core/src/storage/gcs.rs
